requirements.txt
aiohttp
asyncio
discord
requests
pyyaml
